chore: cleanup and use new hooks
